% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cycle.R
\name{add_cell_cycle}
\alias{add_cell_cycle}
\title{Adds cell cycle info to the meta data}
\usage{
add_cell_cycle(mtec, cell_cycle_stage)
}
\arguments{
\item{mtec}{a Seurat object}

\item{cell_cycle_stage}{the cell cycle stage output file from cyclone. Can be
cell cycle stage determined with other methods. Requries that cell names are
row names of the data frame and that these names are identical to the names
in the seurat object.}
}
\description{
This function will add the cell cycle information created by run_cyclone to
the seurat object. Called directly by run_cyclone and works best that way.
}
\keyword{cell}
\keyword{cycle}
\keyword{seurat,}
