% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_processing.R
\name{add_nReads}
\alias{add_nReads}
\title{Adds number of reads to seurat object}
\usage{
add_nReads(mtec, reads_df)
}
\arguments{
\item{mtec}{a seurat object}

\item{reads_df}{the data frame containing number of reads per cell}
}
\description{
This function allows you to add number of reads to seurat metadata
}
\keyword{meta.data,}
\keyword{nReads}
