% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_processing.R
\name{add_perc_mito}
\alias{add_perc_mito}
\title{Adds mito percent to seurat object}
\usage{
add_perc_mito(mtec, mito_pattern = "^mt-")
}
\arguments{
\item{mtec}{a seurat object}

\item{mito_pattern}{OPTIONAL the pattern of mito reads. Defaults to mouse
(^mt-)}
}
\description{
This function allows you to add mito percent to seurat meta data
}
\examples{
add_perc_mito(mTEC.10x.data::mtec_trace_empty)
}
\keyword{meta.data,}
\keyword{perc.mito}
