% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{add_pseudotime}
\alias{add_pseudotime}
\title{Adds the pseudotime information to a Seurat object}
\usage{
add_pseudotime(mtec, pt_obj, method = "monocle", root_state = NULL)
}
\arguments{
\item{mtec}{a Seurat object}

\item{pt_obj}{the pseudotime object}

\item{method}{OPTIONAL the method of pseudotime used. Can be "monocle" or
"dpt". Defaults to "monocle"}

\item{root_state}{OPTIONAL if you want to change the root state for
pseudotime. defaults to NULL.}
}
\description{
This function adds information from a pseudotime object to the seurat meta
data. Some parts of this function were borrowed from
maehrlab/thymusatlastools
}
\examples{
add_pseudotime(mTEC.10x.data::mtec_trace, mTEC.10x.data::mtec_trace_monocle)
add_pseudotime(mTEC.10x.data::mtec_trace, mTEC.10x.data::mtec_trace_dpt,
method = "dpt")
}
\keyword{monocle}
