% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter_seurat.R
\name{change_clus_ids}
\alias{change_clus_ids}
\title{Changes cluster names of seurat object}
\usage{
change_clus_ids(mtec, new_ids = c(1, 2, 6, 0, 4, 3, 5, 7, 8))
}
\arguments{
\item{mtec}{a Seurat object}

\item{new_ids}{OPTIONAL A list of new ids. Must be the same length as current
ids. Defaults to c(1,2,6,0,4,3,5,7,8) as used for the initial mtec object}
}
\description{
This function allows you to change the names of clusters in a seurat object.
This could be to number ordered by developmental or cell type names
}
\examples{
change_clus_ids(mTEC.10x.data::mtec_trace)
change_clus_ids(mTEC.10x.data::mtec_trace,
new_ids = c(8, 7, 6, 5, 4, 3, 2, 1, 0))
}
\keyword{seurat}
