% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{discrete_plots}
\alias{discrete_plots}
\title{Plots discrete colours on a umap, tsne, and or a PCA}
\usage{
discrete_plots(mtec, mtec_data, col_by, UMAP = TRUE, tSNE = FALSE,
  PCA = FALSE, color = NULL, save_plot = NULL, show_legend = TRUE)
}
\arguments{
\item{mtec}{a Seurat object}

\item{mtec_data}{a dataframe containing the values to plot named by
colour_metric}

\item{col_by}{what to use to colour the plot. Can be anything from meta.data
or clusters.}

\item{UMAP}{OPTIONAL if the plot should be a umap. Defaults to TRUE}

\item{tSNE}{OPTIONAL if the plot should be a tSNE. Defaults to FALSE}

\item{PCA}{OPTIONAL if the plot should be a PCA. Defaults to FALSE}

\item{color}{OPTIONAL what color should be used to color the plots. Must be the same
number of items as col_by. Defaults to set1 from RColorBrewer}

\item{save_plot}{OPTIONAL the path of the save file if the file should be saved.
Must end in .pdf or .png. Default is NULL with no saved plot but the plot is printed.}

\item{show_legend}{OPTIONAL if a legend for the plot should be shown. Default
is TRUE}
}
\description{
This function allows you to plot discrete colours on a tsne and or a pca
given a seurat object and dataframe. Must have already run FindClusters by
seurat. Called by tSNE_PCA and works best through that function. Returns
a list with ggplot objects. Returns objects under slots pca, tsne, umap
depending on which of the three you ran.
}
\keyword{PCA,}
\keyword{UMAP,}
\keyword{clusters}
\keyword{tSNE,}
