% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{discretePlots}
\alias{discretePlots}
\title{Plots discrete colours on a umap, tsne, and or a PCA}
\usage{
discretePlots(plot_df, col_by, axis_names = c("dim1", "dim2"),
  color = NULL, save_plot = NULL, show_legend = TRUE)
}
\arguments{
\item{plot_df}{a data_frame containing values used for plotting. This is df
is made by the plotDimRed function. x and y values should be in columns labeled
dim1 and dim2. discrete descriptions of how to color are given in the column
colour_metric.}

\item{col_by}{what to use to colour the plot. Can be anything from meta.data
or clusters. This will be the name of the color bar.}

\item{axis_names}{OPTIONAL What to name the axes. This will be the default
seurat name of the dimensional reduction if called by plotDimRed. Default is
c("dim1", "dim2")}

\item{color}{OPTIONAL what color should be used to color the plots. Must be the same
number of items as col_by. Defaults to set1 from RColorBrewer}

\item{save_plot}{OPTIONAL the path of the save file if the file should be saved.
Must end in .pdf or .png. Default is NULL with no saved plot.}

\item{show_legend}{OPTIONAL if a legend for the plot should be shown. Default
is TRUE}
}
\value{
a ggplot object that is the dimensional reduction of your data
colored by the parameter of your choice.
}
\description{
This function allows you to plot discrete colours on a tsne and or a pca
given a seurat object and dataframe. Must have already run FindClusters by
seurat. Called by plotDimRed and works best through that function. Returns
a ggplot object.
}
\keyword{PCA,}
\keyword{UMAP,}
\keyword{clusters,}
\keyword{dimensional}
\keyword{plot}
\keyword{reduction}
\keyword{tSNE,}
