% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{Exp_plots}
\alias{Exp_plots}
\title{Plots a density plot}
\usage{
Exp_plots(seurat_object, y_val, boxplot = TRUE, vioplot = FALSE,
  density = FALSE)
}
\arguments{
\item{seurat_object}{a Seurat object}

\item{y_val}{the name of a gene or quality metric with which to make the y
axis of the plot}

\item{boxplot}{OPTIONAL if a boxplot should be made. Defaults to TRUE}

\item{vioplot}{OPTIONAL if a violin plot should be made. Defaults to FALSE}

\item{density}{OPTIONAL if a density plot should be made. Defaults to FALSE}
}
\description{
This function allows you to plot density plot given a seurat
object and either a quality metric or a gene.
}
\examples{
Exp_plots(mTEC.10x.data::mtec_trace, "Aire")
Exp_plots(mTEC.10x.data::mtec_trace, "nUMI")
}
