% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{find_markers}
\alias{find_markers}
\title{Finds gene markers between two clusters.}
\usage{
find_markers(seurat_object, ident_1, ident_2, adj_p_val = 0.05,
  logFC = 1, test_use = "wilcox")
}
\arguments{
\item{seurat_object}{A seurat object}

\item{ident_1}{the name of the first cluster you wish to compare}

\item{ident_2}{the name of the second cluster you wish to compare}

\item{adj_p_val}{OPTIONAL the p_value used to call significance.
Defaults to 0.05}

\item{logFC}{OPTIONAL the log fold change used to call significance.
Defaults to 1.0.}

\item{test_use}{OPTIONAL which DE test to use, default is wilcox}
}
\description{
This function finds markers (DE genes) and returns a 
data frame with an additional column containing significance level
(significant/fold change). Meant to be called by significant markers,
but can work independently. Uses FindMarkers from seurat to call
DE genes
}
\keyword{find}
\keyword{markers}
