% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{get_avg_exp}
\alias{get_avg_exp}
\title{Gets average expression of all genes}
\usage{
get_avg_exp(seurat_object, avg_expr_id = "stage")
}
\arguments{
\item{seurat_object}{a seurat object}

\item{avg_expr_id}{the OPTIONAL the id of the meta data column containing
the grouping to separate cells by before calculating average expression.
Default is "stage"}
}
\value{
a data frame of genes by groups with values that are average
expression of each gene within each group.
}
\description{
This function allows you to determine average expression of genes in a 
given grouping. This grouping must be in seurat_object@meta.data
}
\keyword{average}
\keyword{cluster,}
\keyword{expression}
\keyword{gene}
\keyword{group,}
