% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percents_and_counts.R
\name{get_perc_count}
\alias{get_perc_count}
\title{Gets percents and counts from the output of percents_and_counts}
\usage{
get_perc_count(percent_counts_list, list_slot, data_type = "counts")
}
\arguments{
\item{percent_counts_list}{the return value from percents_and_counts}

\item{list_slot}{which slot in the list to pull}

\item{data_type}{OPTIONAL return "counts", "umi", or "percent". Must be one of the
three. Default is "counts"}
}
\value{
a named list of the percent of all genes expressed for the gene set
listed in "list_slot"
}
\description{
This function is a relatively silly work around to get a list of
just the return from the percents or just the return from the counts
for all gene lists. Works best with sapply as shown in the example
}
\examples{
\dontrun{
percents <- sapply(names(percents_counts_all), function(x) 
 get_perc_count(percents_counts_all, x, percent = TRUE), USE.NAMES = TRUE)
}
}
\keyword{count}
\keyword{count,}
\keyword{gene}
\keyword{percent,}
\keyword{set,}
\keyword{umi}
