% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percents_and_counts.R
\name{get_umi}
\alias{get_umi}
\title{Determines median umi count}
\usage{
get_umi(seurat_obj, subset_seurat = FALSE, subset_by = "exp",
  subset_val = "isoControlBeg")
}
\arguments{
\item{seurat_obj}{a Seurat object}

\item{subset_seurat}{OPTIONAL if the seurat object should be subset by some
parameter before determining percent of cycing. Default is FALSE.}

\item{subset_by}{OPTIONAL what meta data column should be used to subset
the data. Default is "exp"}

\item{subset_val}{OPTIONAL What value from the meta data column to use to subset
the seurat object. The value must be a value from the subset_by specified. Default
is "isoControlBeg"}
}
\value{
a value giving the number of unique molecules
}
\description{
This function determines the median umi value given a seurat object. It can
determine the umi value for an entier object or just one subset of it. For
this function, subset_by is best as batches as the output can tell you
differences in median umi count between samples. Helpful for doing downsampling
based on umi
}
\examples{
get_umi(mTEC.10x.data::mtec_trace)
get_umi(mTEC.10x.data::mtec_trace, subset_seurat = TRUE, subset_by = "stage",
subset_val = "Aire_positive")
}
\keyword{umi}
