% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_processing.R
\name{group_cells}
\alias{group_cells}
\title{Groups cells with clustering and/or tSNE}
\usage{
group_cells(mtec, dims_use = 1:10, random_seed = 0, resolution = 0.6,
  cluster = TRUE, tSNE = TRUE, UMAP = TRUE)
}
\arguments{
\item{mtec}{a Seurat object}

\item{dims_use}{OPTIONAL which PC dimensions to use. Should be based on the
heatmap, jackstraw plot, and elbow plot. Defaults to 1:10}

\item{random_seed}{OPTIONAL the random seed to allow reproducibility.
Defaults to 0.}

\item{resolution}{OPTIONAL the resolution for the clustering. Best between
0.6 and 1.2 look at Seurat tutorial for more info. Defaults to 0.6}

\item{cluster}{OPTIONAL if the cells should be clustered. Defaults to TRUE.}

\item{tSNE}{OPTIONAL if a tSNE should be calculated. Defaults to TRUE.}

\item{UMAP}{OPTIONAL if a umap should be calculated. Defaults to TRUE}
}
\description{
This function groups cells by clustering and or by tSNE given a Seurat
object
}
\examples{
group_cells(mTEC.10x.data::mtec_trace)
group_cells(mTEC.10x.data::mtec_trace, dims_use = 1:5, tSNE = FALSE)
}
\keyword{clusters,}
\keyword{tSNE,}
\keyword{umap}
