% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{monocle_genes}
\alias{monocle_genes}
\title{Determines significant genes from a monocle object}
\usage{
monocle_genes(mtec_dev, cores = 1)
}
\arguments{
\item{mtec_dev}{a CellDataSet object}

\item{cores}{OPTIONAL the number of cores to use when running
differentialGeneTest. very slow when run on one core. Defaults to 1.}
}
\description{
This function finds genes that drive pseudotime. While this can be run on a
local, I recommend only running this on a cluster and setting cores to at
least 10. If you only have a local, just use the mtec_trace object as it
already includes the data from monocle.
}
\examples{
\dontrun{
monocle_genes(mTEC.10x.data::mtec_trace_monocle, cores = 10)
}
}
\keyword{monocle}
