% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_process_plots.R
\name{PC_plots}
\alias{PC_plots}
\title{Plots PCA plots, loadings, heatmap, and jackstraw}
\usage{
PC_plots(mtec, visualize_pcs = c(1, 2), test_pcs = 1:12,
  pc_loadings = TRUE, pca_plot = TRUE, heatmap = TRUE,
  jackstraw = FALSE, elbow = TRUE, cells.use = 500,
  use.full = FALSE, num_replicates = 100)
}
\arguments{
\item{mtec}{a Seurat object}

\item{visualize_pcs}{OPTIONAL a list of PCs to plot on a PCA plot and/or
a PC loading plot. If plotting a PCA, the list may only be length 2.
Defaults to c(1, 2)}

\item{test_pcs}{OPTIONAL a range of PCs to test with a heatmap or jackstraw
plot. Defaults to 1:12}

\item{pc_loadings}{OPTIONAL if PC loadings should be visualized. Defaults
to TRUE.}

\item{pca_plot}{OPTIONAL if a PCA plot should be plotted. Defaults to TRUE}

\item{heatmap}{OPTIONAL if a heatmap should be plotted for each of the test
PCs. Defaults to TRUE}

\item{jackstraw}{OPTIONAL if jackstraw should be run to determine PC
significance. May take a long time. Defaults to FALSE.}

\item{elbow}{OPTIONAL if an elbow plot should be created. Defaults to TRUE}

\item{cells.use}{OPTIONAL the number of cells to use to create the heatmap.
Defaults to 500.}

\item{use.full}{OPTIONAL if all genes should be loaded (or only the HVGs).
Defaults to FALSE}

\item{num_replicates}{OPTIONAL the number of replicates to make the jackstraw
plot. Defaults to 100.}
}
\description{
This function allows you to plot a PCA and run all tests to determine
which PCs to use for the remainder of the analysis. More information on
how to determine significant PCs in the Seurat tutorial. Briefly, I used
two methods. 1. plotting the heatmap to see which PCs are driven by genes
that make sense and have clear separation. 2. using jackstraw to determine
which PCs. have an enrichment to genes with low p-values. These p-values
are determined by random permutation of a small subset of the data and
construct a null distribution against which genes can be tested. This 
method is slow.
}
\examples{
PC_plots(mTEC.10x.data::mtec_trace)
\dontrun{
PC_plots(mTEC.10x.data::mtec_trace, visualize_pcs = c(2, 3),
test_pcs = 1:10, jackstraw = TRUE)
}
}
\keyword{PCA}
