% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{PCA_loadings}
\alias{PCA_loadings}
\title{Plots PC loadings}
\usage{
PCA_loadings(seurat_object, PC_val = "PC1", nLoadings = 10,
  left_pull = TRUE, right_pull = TRUE, colour_by = NULL,
  full_loadings = FALSE)
}
\arguments{
\item{seurat_object}{a Seurat object}

\item{PC_val}{OPTIONAL which PC loadings to plot. Defaults to "PC1"}

\item{nLoadings}{OPTIONAL the number of loadings to plot. Defaults to 10}

\item{left_pull}{OPTIONAL plot genes pulling PC to the left. Defaults to
TRUE.}

\item{right_pull}{OPTIONAL plot genes pulling PC to the right. Defaults to
TRUE.}

\item{colour_by}{OPTIONAL a list of genes used for colouring. Defaults to
NULL}

\item{full_loadings}{OPTIONAL if all loadings should be used. If false, only
hvgs used
to create the PCA will be plotted. Defaults to FALSE.}
}
\description{
This function allows you to plot PCA loadings for any number of genes for any
PC given a Seurat object
}
\examples{
PCA_loadings(mTEC.10x.data::mtec_trace)
PCA_loadings(mTEC.10x.data::mtec_trace, PC_val = "PC2", nLoadings = 20,
			colour_by = mTEC.10x.data::tuft_markers, left_pull = FALSE)
}
\keyword{PCA,}
\keyword{clusters}
\keyword{tSNE,}
