% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percents_and_counts.R
\name{percent_cycling_cells}
\alias{percent_cycling_cells}
\title{Determines percent of cycling cells}
\usage{
percent_cycling_cells(seurat_obj, subset_seurat = FALSE,
  subset_by = "exp", subset_val = "isoControlBeg")
}
\arguments{
\item{seurat_obj}{a Seurat object}

\item{subset_seurat}{OPTIONAL if the seurat object should be subset by some
parameter before determining percent of cycing. Default is FALSE.}

\item{subset_by}{OPTIONAL what meta data column should be used to subset
the data. Default is "exp"}

\item{subset_val}{OPTIONAL What value from the meta data column to use to subset
the seurat object. The value must be a value from the subset_by specified. Default
is "isoControlBeg"}
}
\value{
a named list containing the percent of cycling cells and named by subset_val
If no data set was provided, the list will be named "all"
}
\description{
This function determines the percent of cycling cells. The input must
be a seurat object with a meta_data column called "cycle_phase" and containing
values "G2M", "G1", "S". Returns a named list with the percent of cells that
are classified as G2M or S.
}
\examples{
percent_cycling_cells(mTEC.10x.data::mtec_trace)
}
\keyword{cell}
\keyword{cycle,}
\keyword{percent}
