% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percents_and_counts.R
\name{percent_list}
\alias{percent_list}
\title{Returns the percent of genes expressed in a data set}
\usage{
percent_list(cell_matrix, gene_list, gene_list_name)
}
\arguments{
\item{cell_matrix}{a matrix with cells as columns and genes as rows}

\item{gene_lists}{a list of genes}

\item{list}{name the name of the gene list}
}
\value{
a named list of the percent of all genes expressed. The list is named by
the name of the gene set
}
\description{
This function determines the percent of genes in a gene set represented
in the data set. This means the percent of genes from the gene list that
is expressed in at least one cell. Called by percents_and_counts, but can
work on its own too
}
\keyword{gene}
\keyword{percent,}
\keyword{set}
