% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percents_and_counts.R
\name{percents_and_counts}
\alias{percents_and_counts}
\title{Returns information about gene sets}
\usage{
percents_and_counts(seurat_obj, gene_lists, downsample_UMI = FALSE,
  one_batch = FALSE, batch = "exp", batch_name = "all_cells",
  lowest_UMI = NULL, count = "genes")
}
\arguments{
\item{seurat_obj}{a Seurat object}

\item{gene_lists}{a named list of gene sets. Can be just one gene set or many.}

\item{downsample_UMI}{OPTIONAL if the raw data should be downsampled to a 
specific value. Default is FALSE}

\item{one_batch}{OPTIONAL if the seurat object should be subset. This is
especially helpful when comparing between samples within one seurat object.
Default it FALSE.}

\item{batch}{OPTIONAL if one_batch is true, what should be used to downsample?
Default is "exp", but must be a column of metadata}

\item{batch_name}{OPTIONAL the group to keep. This must be a value in the 
metadata column specified by batch. This is also the name of the returned
list. Default is "all_cells".}

\item{lowest_UMI}{OPTIONAL The value used to downsample the UMI. Default is NULL}

\item{count}{OPTIONAL if the count of genes or UMI is returned. Must be "genes" or
"UMI". Default is "genes"}
}
\value{
a list with two parts. One part is the count of genes/UMI the second is
the percent of all genes expressed. Both parts contain lists of every gene set
named by the gene set
}
\description{
This function determines the number of genes or UMIs for genes in each gene set
expressed in each cell and also determines the percent of genes in each gene
set represented in the data set. This means the percent of genes from the gene
list that is expressed in at least one cell.
}
\keyword{gene,}
\keyword{percent}
\keyword{umi,}
