% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot_avg_exp_genes}
\alias{plot_avg_exp_genes}
\title{Plots average expression of a gene in a population in a line plot}
\usage{
plot_avg_exp_genes(seurat_object, gene_list, save_plot = NULL, ...)
}
\arguments{
\item{seurat_object}{a Seurat object}

\item{save_plot}{OPTIONAL the name of the plot to be saved. If no value
is given, the plot won't be saved and will just be returned. Default
is NULL.}

\item{gene_set}{a list of genes to plot expression in a group of cells
the group can be determined using the parameters for get_avg_exp}
}
\value{
a ggplot object that is the line plot of average expression of
each gene in the gene list in each group.
}
\description{
This function allows you to plot average expression of a set of genes
in populations of cells within your data set. The populations can
be changed using the parameters for get_avg_exp.
}
\examples{
plot_avg_exp(mTEC.10x.data::mtec_trace, gene_list = c("Aire", "Fezf2"))
}
\keyword{average}
\keyword{expression,}
\keyword{gene}
\keyword{line}
\keyword{plot}
