% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot_gene_set}
\alias{plot_gene_set}
\title{Plots average expression of a set of genes as colors on a dim red plot}
\usage{
plot_gene_set(seurat_obj, gene_set, plot_name, return_plot = TRUE, ...)
}
\arguments{
\item{gene_set}{a list of genes to compute the average expression for}

\item{plot_name}{what to name the plot}

\item{return_plot}{OPTIONAL if the average expression should be plotted.
This calls the plotDimRed and inherits all of those plotting parameters.
If return_plot is true, the ggplot object will be returned. If return_plot
is false, the seurat object will be returned with the average expression
values in the meta.data. Default is TRUE.}

\item{seurat_object}{a Seurat object}
}
\value{
either a ggplot object that is the dimensional reduction of your data
colored by the average expression of your genes or a seurat object updated to
include average expression values for each cell. Return type is decided by
return_plot.
}
\description{
This function allows you to plot average expression of a group of genes on
' a umap, tsne, cca, or a pca given a seurat object Returns
a list with ggplot objects. Returns objects under slots pca, tsne, umap
depending on which of the three you ran.
}
\examples{
plot_gene_set(mTEC.10x.data::mtec_trace, gene_set = c("Aire", "Fezf2"),
  plot_name = "mtecGenes")
plot_gene_set(mTEC.10x.data::mtec_trace, gene_set = c("Aire", "Fezf2"),
  plot_name = "mtecGenes", plot_type = "tsne")
}
\keyword{average}
\keyword{dim}
\keyword{expression,}
\keyword{gene}
\keyword{plot}
\keyword{reduction}
