% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plots a heatmap and will label genes of interest if given. Can only plot
DE genes put in the DE slot after running significant markers.}
\usage{
plot_heatmap(mtec, cell_color = NULL, subset_list = NULL,
  color_list = NULL, color_list2 = NULL, order_cells = TRUE,
  seed = 0)
}
\arguments{
\item{mtec}{A seurat object}

\item{cell_color}{OPTIONAL a list of colors to color the cell labels at the
top of the plot. Defaults to NULL and uses color brewer set1}

\item{subset_list}{OPTIONAL a list of genes to subset the data. Defaults
to NULL}

\item{color_list}{OPTIONAL a list of genes to color gene names on the plot.
Defaults to NULL}

\item{order_cells}{OPTIONAL if cells should be ordered based on cluster number.
Defaults to TRUE}

\item{seed}{OPTIONAL the seed to use for reproducability. Defaults to 0.}
}
\description{
Plots a heatmap and will label genes of interest if given. Can only plot
DE genes put in the DE slot after running significant markers.
}
\examples{
\dontrun{
plot_heatmap(mTEC.10x.data::mtec_trace)
}
plot_heatmap(mTEC.10x.data::mtec_trace, subset_list = mTEC.10x.data::TFs,
color_list = c("Aire", "Hmgb2"), order_cells = TRUE)
}
\keyword{heatmap}
