% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plots a volcano plot
This function plots a volcano plot and labels highest fold change
genes (both up and down). Works best when called by significant_markers,
but will work if given a full gene list.}
\usage{
plot_volcano(marker_genes, title, num_genes = 5)
}
\arguments{
\item{marker_genes}{a data frame containing a list of genes and
a colum containing significance information. This df is made by the 
find_markers function}

\item{title}{the title of the plot}

\item{num_genes}{OPTIONAL the number of genes to label. Defalts to
5.}
}
\description{
Plots a volcano plot
This function plots a volcano plot and labels highest fold change
genes (both up and down). Works best when called by significant_markers,
but will work if given a full gene list.
}
\keyword{plot}
\keyword{volcano}
