% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plotDimRed}
\alias{plotDimRed}
\title{Plots a UMAP and/or tSNE and/or a PCA}
\usage{
plotDimRed(mtec, col_by, plot_type = "umap", dims_use = NULL, ...)
}
\arguments{
\item{mtec}{a Seurat object}

\item{col_by}{the name of a gene or column from meta data with which to
colour the plot. Can also be "cluster"}

\item{plot_type}{OPTIONAL what type of diminsional reduction to plot. Must be 
a diminisional reduction listed in the dr slot of the seurat object. Default is
UMAP.}

\item{dims_use}{OPTIONAL}
}
\value{
a ggplot object that is the dimensional reduction of your data
colored by the parameter of your choice.
}
\description{
This function allows you to plot a tsne and/or a pca given a seurat
object and either a quality metric, cluster, or a gene. Can only plot genes,
anything from meta.data (quality) or cluster. Returns
a list with ggplot objects. Returns objects under slots pca, tsne, umap
depending on which of the three you ran.
}
\examples{
plotDimRed(mTEC.10x.data::mtec_trace, "Aire")
plotDimRed(mTEC.10x.data::mtec_trace, "nUMI", tSNE = FALSE, PCA = TRUE)
}
\keyword{PCA,}
\keyword{UMAP,}
\keyword{diminsional}
\keyword{plot}
\keyword{reduction}
\keyword{tSNE,}
