% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_funcs.R
\name{populations_dfs}
\alias{populations_dfs}
\title{Makes a dataframe from multiple seurat objects for one variable}
\usage{
populations_dfs(seurat_object, sample_name, stage_df_all)
}
\arguments{
\item{seurat_object}{a Seurat object}

\item{sample_name}{the name of one sample to add to the plot}

\item{stage_df_all}{the data frame containing all samples}
}
\description{
This function allows you to take multiple seurat objects and combine
them for one variable (like cell type) to make a plot of population percents
Could be improved to make the entire df and call the plotting function
}
\keyword{percents}
\keyword{population}
