% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{populationsDFS}
\alias{populationsDFS}
\title{Makes a dataframe from multiple seurat objects for one variable}
\usage{
populationsDFS(seurat_object, sample_name, subsample = FALSE,
  subsample_by = "exp", meta_data_col = "stage")
}
\arguments{
\item{seurat_object}{a Seurat object}

\item{sample_name}{the name of one sample to add to the plot}

\item{subsample}{OPTIONAL if the seurat object should be subsampled by some
parameter. For example, if you wanted to determine the percent of a cell type
in all experiments, subsample would be false. If you wanted to determine the
percent of a cell type in each experiment, this would be TRUE. Default is FALSE
subsample_by OPTIONAL what to use to subsample. Must be a column of meta data.
Default is exp.
meta_data_col OPTIONAL which meta data column to use to determine cell numbers.
For example, the meta data column that describes cell type. Default is "stage"}
}
\description{
This function allows you to take multiple seurat objects and combine
them for one variable (like cell type) to make a plot of population percents
Could be improved to make the entire df and call the plotting function
}
\keyword{percents}
\keyword{population}
