% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_processing.R
\name{process_cells}
\alias{process_cells}
\title{Initially processes cells in seurat object}
\usage{
process_cells(mtec, filter_by = c("nGene", "percent_mito"),
  low_thresholds = c(200, -Inf), high_thresholds = c(7500, 0.1),
  scale_factor = 10000, filter_cells = TRUE, normalize_data = TRUE,
  find_var_genes = TRUE, scale_data = TRUE, PCA = TRUE)
}
\arguments{
\item{mtec}{a seurat object}

\item{filter_by}{OPTIONAL what to use to filter cells. Must be in object
meta.data. Can be a list. Defaults to c("nGene", "percent.mito")}

\item{low_thresholds}{OPTIONAL low threshold for each factor used to filter
cells. Can be a list. Defaults to c(200, -Inf)}

\item{high_thresholds}{OPTIONAL high threshold for each factor used to filter
cells. Can be a list. Defaults to c(7500, 0.1)}

\item{scale_factor}{OPTIONAL how to scale the data to normalize. Defaults to
10000.}

\item{filter_cells}{OPTIONAL if cells should be filtered. Defaults to TRUE}

\item{normalize_data}{OPTIONAL if values should be normalized. Defaults to
TRUE.}

\item{find_var_genes}{OPTIONAL if variable genes should be determined.
Defaults to TRUE.}

\item{scale_data}{OPTIONAL if the data should be scaled. Defaults to TRUE}

\item{PCA}{OPTIONAL if PCs should be found. Defaults to TRUE.}
}
\description{
This function allows you to filter cells, normalize data,
find variable genes,and scale data
}
\examples{
process_cells(mTEC.10x.data::mtec_trace_empty, filter_by = c("nGene"),
high_thresholds = c(7500), low_thresholds = c(200))
}
\keyword{meta.data,}
\keyword{perc.mito}
