% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_process_plots.R
\name{qc_plot}
\alias{qc_plot}
\title{Plots qc information}
\usage{
qc_plot(mtec, qc_vals = c("nGene", "nUMI", "percent_mito"))
}
\arguments{
\item{mtec}{a Seurat object}

\item{qc_vals}{OPTIONAL a list of qc information to plot. Defaults to
c("nGene", "nUMI", "percent.mito").}
}
\description{
This function allows you to plot a qc information given a seurat
object.
}
\examples{
qc_plot(mTEC.10x.data::mtec_trace, qc_vals = c("nGene", "nUMI"))
}
\keyword{qc}
