% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cycle.R
\name{run_cyclone}
\alias{run_cyclone}
\title{Runs cell cycle analysis on cells}
\usage{
run_cyclone(mtec, conversion_file, seed = 0, add_to_obj = TRUE)
}
\arguments{
\item{mtec}{a Seurat object}

\item{conversion_file}{a file to use to convert gene names to ensembl ids.}

\item{seed}{OPTIONAL seed for reproducibility. Defaults to 0}

\item{add_to_obj}{OPTIONAL if the data should be added to the Seurat object.
Defaults to TRUE}
}
\description{
This function allows you to run cell cycle analysis on the population of
cells using cyclone from scran.
}
\examples{
\dontrun{
run_cyclone(mTEC.10x.data::mtec_trace)
run_cyclone(mTEC.10x.data::mtec_trace, seed = 10)
}
}
\keyword{scran}
