% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{run_monocle}
\alias{run_monocle}
\title{Runs monocle on cells}
\usage{
run_monocle(mtec, quality_plots = FALSE, cores = 1, seed = 0)
}
\arguments{
\item{mtec}{a Seurat object}

\item{quality_plots}{OPTIONAL if quality plots (including density plot of
expression values, PC variance explained, and tSNE coloured by monocle and
seurat clusters) should be plotted. Defaults to FALSE.}

\item{cores}{OPTIONAL the number of cores to use when running
differentialGeneTest. very slow when run on one core. Defaults to 1.}

\item{seed}{OPTIONAL seed for reproducibility. Defaults to 0}
}
\description{
This function allows you to run monocle on your samples. It also will add the
monocle output into the seurat object. While this can be run on a local, I
recommend only running this on a cluster and setting cores to at least 10.
If you only have a local, just use the mtec_trace object as it already
includes the data from monocle.
}
\examples{
\dontrun{
run_monocle(mTEC.10x.data::mtec_trace, cores = 10)
}
}
\keyword{monocle}
