% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter_seurat.R
\name{set_stage}
\alias{set_stage}
\title{Adds stage information to seurat object}
\usage{
set_stage(mtec_obj, stage_list)
}
\arguments{
\item{mtec_obj}{a Seurat object}

\item{stage_list}{a list of stages and their corresponding cluster numbers.}
}
\description{
This function allows you to add stage information to a seurat object. 
It depends on a list input by the user that links cluster number to 
stage.
}
\examples{
stage_list <- c("0" = "cTEC", "1" = "Immature", "2" = "Immature",
               "3" = "Immature", "4" = "Intermediate",
               "5" = "Mature", "6" = "Late_mature",
               "7" = "Tuft")
mtec_trace <- set_stage(mTEC.10x.data::mtec_trace, stage_list)
}
\keyword{clusters}
\keyword{seurat,}
