% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{significant_markers}
\alias{significant_markers}
\title{Finds significant markers (DE genes) between clusters}
\usage{
significant_markers(mtec, cluster_list = NULL, adj_p_val = 0.05,
  logFC = 1, plot_volcano = FALSE, num_genes = 5,
  test_use = "wilcox")
}
\arguments{
\item{mtec}{A seurat object}

\item{cluster_list}{OPTIONAL A list of cluster names to perform DE
between. Defaults to NULL (all clusters in object)}

\item{adj_p_val}{OPTIONAL The pvalue to call for significance
Defaults to 0.05}

\item{logFC}{OPTIONAL The log full change to call significant.
Defaults to 1 (a 2 fold change)}

\item{plot_volcano}{OPTIONAL if a volcano plot should be made.
Defaults to FALSE.}

\item{num_genes}{OPTIONAL the number of genes to plot on the 
volcano plot if you choose to plot it. Defaults to 5.}

\item{test_use}{OPTIONAL which DE test to use}
}
\description{
This function allows you to find significant markers between
two clusters (or pairwise between any clusters you choose).
Uses seurat FindMarkers to find DE genes.
Requries a mTEC seurat object where FindClusters has already been
run. If no list is given, will perform pairwise analysis of all
clusters. If list is given, will perform pairwise analysis of list
given. Can optionally plot a volcano plot with any number of top
genes labeled on the plot. Returns a new seurat object containing
DE genes in the assay slot (in a list called DE)
}
\examples{
\dontrun{
significant_markers(mTEC.10x.data::mtec_trace)
significant_markers(mTEC.10x.data::mtec_trace, cluster_list = c(5, 6),
plot_volcano = TRUE)
}
}
\keyword{DE,}
\keyword{plot}
\keyword{volcano}
