% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter_seurat.R
\name{subset_seurat}
\alias{subset_seurat}
\title{Subsets the seurat object to selected clusters}
\usage{
subset_seurat(mtec, clusters)
}
\arguments{
\item{mtec}{a Seurat object}

\item{clusters}{A list clusters to subset (ex c(1, 2, 3))}
}
\description{
This function allows you to generate an empty seurat object with only
the cells in clusters of interest. Then the processing pipeline can
be repeated with the different group of cells.
}
\examples{
subset_seurat(mTEC.10x.data::mtec_trace, clusters = c(4, 5, 6))
}
\keyword{clusters}
\keyword{seurat,}
