% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{trio_plots}
\alias{trio_plots}
\title{Plots three plots in one image}
\usage{
trio_plots(seurat_object, geneset, cell_cycle = FALSE,
  plot_jitter = TRUE, plot_violin = FALSE, jitter_and_violin = FALSE,
  color = NULL, sep_by = "cluster", save_plot = NULL, nrow = NULL,
  ncol = NULL, group_color = TRUE)
}
\arguments{
\item{seurat_object}{a Seurat object}

\item{geneset}{a list of three genes to plot}

\item{cell_cycle}{OPTIONAL if to color plots based on cell cycle scores.
Defaults to FALSE}

\item{plot_jitter}{OPTIONAL if the plot should be a jitter plot. Defaults to TRUE}

\item{plot_violin}{OPTIONAL if the plot should be a violin. Defaults to FALSE}

\item{jitter_and_violin}{OPTIONAL if the plot should be a jitter on top of a violin.
Defaults to FALSE}

\item{color}{OPTIONAL what colors should be used to color the plots. Defaults to 
RColorBrewer set 9.}

\item{sep_by}{OPTIONAL what to use to separate on the x axis. Defaults to cluster}

\item{save_plot}{OPTIONAL the path of the save file if the file should be saved.
Must end in .pdf or .png. Default is NULL with no saved plot but the plot is printed.}

\item{nrow}{OPTIONAL number of rows for output plots default is number of plots}

\item{ncol}{OPTIONAL number of columns for output plots default is 1}

\item{group_color}{OPTIONAL if a plot of group colors should be output. Default
is true}
}
\value{
a ggplot object containing the jitter or violin plots arranged with
grid extra
}
\description{
This function allows you to plot three plots on one image that are violin
or jitter plots given a seurat object. Must have already run FindClusters by
seurat.
}
\keyword{jitter,}
\keyword{trio}
\keyword{violin,}
