% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{tSNE_PCA}
\alias{tSNE_PCA}
\title{Plots a UMAP and/or tSNE and/or a PCA}
\usage{
tSNE_PCA(mtec, col_by, UMAP = TRUE, tSNE = FALSE, PCA = FALSE,
  color = NULL, save_plot = NULL, show_legend = TRUE)
}
\arguments{
\item{mtec}{a Seurat object}

\item{col_by}{the name of a gene or column from meta data with which to
colour the plot. Can also be "cluster"}

\item{UMAP}{OPTIONAL if the plot should be a umap. Defaults to TRUE}

\item{tSNE}{OPTIONAL if the plot should be a tSNE. Defaults to FALSE}

\item{PCA}{OPTIONAL if the plot should be a PCA. Defaults to FALSE}

\item{color}{OPTIONAL the colors used to color the plots. For gene values, provide a list
of two colors. For discrete values, provide a list as long as the object you want to color
example: if you are coloring clusters and have 5 clusters, provide 5 colors. Default is
set1 from ggplot2 for discrete and red/blue for continuous.}

\item{save_plot}{OPTIONAL the path of the save file if the file should be saved.
Must end in .pdf or .png. Default is NULL with no saved plot but the plot is printed.}

\item{show_legend}{OPTIONAL if a legend for the plot should be shown. Default
is TRUE}
}
\description{
This function allows you to plot a tsne and/or a pca given a seurat
object and either a quality metric, cluster, or a gene. Can only plot genes,
anything from meta.data (quality) or cluster. Returns
a list with ggplot objects. Returns objects under slots pca, tsne, umap
depending on which of the three you ran.
}
\examples{
tSNE_PCA(mTEC.10x.data::mtec_trace, "Aire")
tSNE_PCA(mTEC.10x.data::mtec_trace, "nUMI", tSNE = FALSE, PCA = TRUE)
}
\keyword{PCA,}
\keyword{UMAP}
\keyword{tSNE,}
