% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percents_and_counts.R
\name{umi_count_function}
\alias{umi_count_function}
\title{Returns the number of UMIs from a gene set in each cell}
\usage{
umi_count_function(cell_matrix, gene_list, list_name)
}
\arguments{
\item{cell_matrix}{a matrix with cells as columns and genes as rows}

\item{gene_lists}{a list of genes}

\item{list}{name the name of the gene list}
}
\value{
a named list of the number of UMIs expressed in every cell
}
\description{
This function determines the number of UMIs for genes in a gene set
expressed in each cell. Called by percents_and_counts, but can
work on its own too
}
\keyword{count,}
\keyword{gene}
\keyword{set}
\keyword{umi}
